@@ -1,16 +1,19 @@
 {
 	"name": "wrap-ansi",
-	"version": "4.0.0",
+	"version": "8.0.1",
 	"description": "Wordwrap a string with ANSI escape codes",
 	"license": "MIT",
 	"repository": "chalk/wrap-ansi",
+	"funding": "https://github.com/chalk/wrap-ansi?sponsor=1",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=6"
+		"node": ">=12"
 	},
 	"scripts": {
 		"test": "xo && nyc ava"
@@ -46,17 +49,16 @@
 		"text"
 	],
 	"dependencies": {
-		"ansi-styles": "^3.2.0",
-		"string-width": "^2.1.1",
-		"strip-ansi": "^4.0.0"
+		"ansi-styles": "^6.1.0",
+		"string-width": "^5.0.1",
+		"strip-ansi": "^7.0.1"
 	},
 	"devDependencies": {
-		"ava": "^0.25.0",
-		"chalk": "^2.0.1",
-		"coveralls": "^3.0.0",
-		"has-ansi": "^3.0.0",
-		"nyc": "^13.0.1",
-		"strip-ansi": "^4.0.0",
-		"xo": "^0.22.0"
+		"ava": "^3.15.0",
+		"chalk": "^4.1.2",
+		"coveralls": "^3.1.1",
+		"has-ansi": "^5.0.1",
+		"nyc": "^15.1.0",
+		"xo": "^0.44.0"
 	}
 }
