@@ -1,16 +1,21 @@
-'use strict';
-const stringWidth = require('string-width');
-const stripAnsi = require('strip-ansi');
-const ansiStyles = require('ansi-styles');
+import stringWidth from 'string-width';
+import stripAnsi from 'strip-ansi';
+import ansiStyles from 'ansi-styles';
 
 const ESCAPES = new Set([
 	'\u001B',
-	'\u009B'
+	'\u009B',
 ]);
 
 const END_CODE = 39;
+const ANSI_ESCAPE_BELL = '\u0007';
+const ANSI_CSI = '[';
+const ANSI_OSC = ']';
+const ANSI_SGR_TERMINATOR = 'm';
+const ANSI_ESCAPE_LINK = `${ANSI_OSC}8;;`;
 
-const wrapAnsi = code => `${ESCAPES.values().next().value}[${code}m`;
+const wrapAnsiCode = code => `${ESCAPES.values().next().value}${ANSI_CSI}${code}${ANSI_SGR_TERMINATOR}`;
+const wrapAnsiHyperlink = uri => `${ESCAPES.values().next().value}${ANSI_ESCAPE_LINK}${uri}${ANSI_ESCAPE_BELL}`;
 
 // Calculate the length of words split on ' ', ignoring
 // the extra characters added by ansi escape codes
@@ -21,7 +26,8 @@ const wordLengths = string => string.split(' ').map(character => stringWidth(cha
 const wrapWord = (rows, word, columns) => {
 	const characters = [...word];
 
-	let insideEscape = false;
+	let isInsideEscape = false;
+	let isInsideLinkEscape = false;
 	let visible = stringWidth(stripAnsi(rows[rows.length - 1]));
 
 	for (const [index, character] of characters.entries()) {
@@ -35,13 +41,20 @@ const wrapWord = (rows, word, columns) => {
 		}
 
 		if (ESCAPES.has(character)) {
-			insideEscape = true;
-		} else if (insideEscape && character === 'm') {
-			insideEscape = false;
-			continue;
+			isInsideEscape = true;
+			isInsideLinkEscape = characters.slice(index + 1).join('').startsWith(ANSI_ESCAPE_LINK);
 		}
 
-		if (insideEscape) {
+		if (isInsideEscape) {
+			if (isInsideLinkEscape) {
+				if (character === ANSI_ESCAPE_BELL) {
+					isInsideEscape = false;
+					isInsideLinkEscape = false;
+				}
+			} else if (character === ANSI_SGR_TERMINATOR) {
+				isInsideEscape = false;
+			}
+
 			continue;
 		}
 
@@ -60,51 +73,77 @@ const wrapWord = (rows, word, columns) => {
 	}
 };
 
-// The wrap-ansi module can be invoked
-// in either 'hard' or 'soft' wrap mode
+// Trims spaces from a string ignoring invisible sequences
+const stringVisibleTrimSpacesRight = string => {
+	const words = string.split(' ');
+	let last = words.length;
+
+	while (last > 0) {
+		if (stringWidth(words[last - 1]) > 0) {
+			break;
+		}
+
+		last--;
+	}
+
+	if (last === words.length) {
+		return string;
+	}
+
+	return words.slice(0, last).join(' ') + words.slice(last).join('');
+};
+
+// The wrap-ansi module can be invoked in either 'hard' or 'soft' wrap mode
 //
-// 'hard' will never allow a string to take up more
-// than columns characters
+// 'hard' will never allow a string to take up more than columns characters
 //
 // 'soft' allows long words to expand past the column length
 const exec = (string, columns, options = {}) => {
-	if (string.trim() === '') {
-		return options.trim === false ? string : string.trim();
+	if (options.trim !== false && string.trim() === '') {
+		return '';
 	}
 
-	let pre = '';
-	let ret = '';
+	let returnValue = '';
 	let escapeCode;
+	let escapeUrl;
 
 	const lengths = wordLengths(string);
-	const rows = [''];
+	let rows = [''];
 
 	for (const [index, word] of string.split(' ').entries()) {
-		rows[rows.length - 1] = options.trim === false ? rows[rows.length - 1] : rows[rows.length - 1].trim();
+		if (options.trim !== false) {
+			rows[rows.length - 1] = rows[rows.length - 1].trimStart();
+		}
+
 		let rowLength = stringWidth(rows[rows.length - 1]);
 
-		if (rowLength || word === '') {
-			if (rowLength === columns && options.wordWrap === false) {
+		if (index !== 0) {
+			if (rowLength >= columns && (options.wordWrap === false || options.trim === false)) {
 				// If we start with a new word but the current row length equals the length of the columns, add a new row
 				rows.push('');
 				rowLength = 0;
 			}
 
-			rows[rows.length - 1] += ' ';
-			rowLength++;
+			if (rowLength > 0 || options.trim === false) {
+				rows[rows.length - 1] += ' ';
+				rowLength++;
+			}
 		}
 
-		// In 'hard' wrap mode, the length of a line is
-		// never allowed to extend past 'columns'
-		if (lengths[index] > columns && options.hard) {
-			if (rowLength) {
+		// In 'hard' wrap mode, the length of a line is never allowed to extend past 'columns'
+		if (options.hard && lengths[index] > columns) {
+			const remainingColumns = (columns - rowLength);
+			const breaksStartingThisLine = 1 + Math.floor((lengths[index] - remainingColumns - 1) / columns);
+			const breaksStartingNextLine = Math.floor((lengths[index] - 1) / columns);
+			if (breaksStartingNextLine < breaksStartingThisLine) {
 				rows.push('');
 			}
+
 			wrapWord(rows, word, columns);
 			continue;
 		}
 
-		if (rowLength + lengths[index] > columns && rowLength > 0) {
+		if (rowLength + lengths[index] > columns && rowLength > 0 && lengths[index] > 0) {
 			if (options.wordWrap === false && rowLength < columns) {
 				wrapWord(rows, word, columns);
 				continue;
@@ -121,35 +160,55 @@ const exec = (string, columns, options = {}) => {
 		rows[rows.length - 1] += word;
 	}
 
-	pre = rows.map(row => options.trim === false ? row : row.trim()).join('\n');
+	if (options.trim !== false) {
+		rows = rows.map(row => stringVisibleTrimSpacesRight(row));
+	}
 
-	for (const [index, character] of [...pre].entries()) {
-		ret += character;
+	const pre = [...rows.join('\n')];
+
+	for (const [index, character] of pre.entries()) {
+		returnValue += character;
 
 		if (ESCAPES.has(character)) {
-			const code = parseFloat(/\d[^m]*/.exec(pre.slice(index, index + 4)));
-			escapeCode = code === END_CODE ? null : code;
+			const {groups} = new RegExp(`(?:\\${ANSI_CSI}(?<code>\\d+)m|\\${ANSI_ESCAPE_LINK}(?<uri>.*)${ANSI_ESCAPE_BELL})`).exec(pre.slice(index).join('')) || {groups: {}};
+			if (groups.code !== undefined) {
+				const code = Number.parseFloat(groups.code);
+				escapeCode = code === END_CODE ? undefined : code;
+			} else if (groups.uri !== undefined) {
+				escapeUrl = groups.uri.length === 0 ? undefined : groups.uri;
+			}
 		}
 
 		const code = ansiStyles.codes.get(Number(escapeCode));
 
-		if (escapeCode && code) {
-			if (pre[index + 1] === '\n') {
-				ret += wrapAnsi(code);
-			} else if (character === '\n') {
-				ret += wrapAnsi(escapeCode);
+		if (pre[index + 1] === '\n') {
+			if (escapeUrl) {
+				returnValue += wrapAnsiHyperlink('');
+			}
+
+			if (escapeCode && code) {
+				returnValue += wrapAnsiCode(code);
+			}
+		} else if (character === '\n') {
+			if (escapeCode && code) {
+				returnValue += wrapAnsiCode(escapeCode);
+			}
+
+			if (escapeUrl) {
+				returnValue += wrapAnsiHyperlink(escapeUrl);
 			}
 		}
 	}
 
-	return ret;
+	return returnValue;
 };
 
 // For each newline, invoke the method separately
-module.exports = (string, columns, options) => {
+export default function wrapAnsi(string, columns, options) {
 	return String(string)
 		.normalize()
+		.replace(/\r\n/g, '\n')
 		.split('\n')
 		.map(line => exec(line, columns, options))
 		.join('\n');
-};
+}
